Menambahkan fitur alert di user management

+++ src/components/dashboard/UserManagementComponent.jsx
@@ -5,7 +5,7 @@ import ComponentTable from '../component/Table';
 // import AddUserForm from '../component/UserForm'; // Import komponen form
 import { useEffect, useState } from 'react';
 import UserFormWithMap from '../component/UserFormWithMap';
-// import Swal from 'sweetalert2';
+import Swal from 'sweetalert2';
 
 const UserManagement = () => {
     const headers = ["name", "email", "phone_number", "location"];
@@ -28,7 +28,7 @@ const UserManagement = () => {
     const handleAddUserClick = () => {
         setIsAddingUser(true); // Set state untuk menampilkan form
         setIsEditingUser(false);
-        setSelectedUser(null);
+        setSelectedUser({});
         // console.log(isAddingUser);
     };
 
@@ -44,26 +44,56 @@ const UserManagement = () => {
         setIsAddingUser(false); // Kembali ke tampilan tabel setelah submit
         setIsEditingUser(false);
         setSelectedUser(null);
-        fetchUsers(currentPage); // Refresh data pengguna
+
+        fetchUsers(currentPage) // Refresh data pengguna
+            .then(() => {
+                const message = isEditingUser ? 'User data successfully updated.' : 'User successfully added';
+                Swal.fire({
+                    title: 'Success!',
+                    text: message,
+                    icon: 'success',
+                    confirmButtonText: 'OK'
+                });
+            })
+            .catch((error) => {
+                Swal.fire({
+                    title: 'Error!',
+                    text: error.message,
+                    icon: 'error',
+                    confirmButtonText: 'OK'
+                });
+            });
     };
 
     const handleDeleteUserClick = async (userId) => {
-        try {
+        const result = await Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+        });
 
-            const response = await fetch(`http://localhost:3000/users/${ userId }`, {
-                method: 'DELETE',
-            });
+        if (result.isConfirmed) {
+            try {
+                const response = await fetch(`http://localhost:3000/users/${ userId }`, {
+                    method: 'DELETE',
+                });
 
-            if (!response.ok) {
-                throw new Error('Failed to delete user.');
+                if (!response.ok) {
+                    throw new Error('Failed to delete user.');
+                }
+
+                Swal.fire('Deleted!', 'User has been deleted.', 'success');
+                fetchUsers(currentPage);
+            } catch (error) {
+                console.error(error);
+                Swal.fire('Error!', 'Failed to delete user.', 'error');
             }
-
-
-
-            fetchUsers(currentPage); // Refresh data pengguna setelah delete
-        } catch (error) {
-            console.error(error);
         }
+
     };
     return (
         <div className="flex flex-col p-4 h-full overflow-y-auto ">
